feat: Introduce observation and reward calculation utilities for enhanced LudoGym environment
