+++ ludo_rl/envs/ludo_env.py
@@ -22,6 +22,7 @@ Future extensions: maskable action space, multi-agent self-play environment.
 
 from __future__ import annotations
 
+import math
 import random
 from typing import Any, Dict, List, Optional
 
@@ -34,7 +35,11 @@ from ludo.game import LudoGame
 from ludo.player import Player, PlayerColor
 from ludo.strategy import StrategyFactory
 
+from .builders.observation_builder import ObservationBuilder
+from .calculators.reward_calculator import RewardCalculator
 from .model import EnvConfig
+from .simulators.opponent_simulator import OpponentSimulator
+from .utils.move_utils import MoveUtils
 
 # --------------------------------------------------------------------------------------
 # Environment
@@ -79,7 +84,7 @@ class LudoGymEnv(gym.Env):
                     pass
 
         # Spaces
-        obs_dim = self._compute_observation_size()
+        obs_dim = self.obs_builder._compute_observation_size()
         self.observation_space = spaces.Box(
             low=-1.0, high=1.0, shape=(obs_dim,), dtype=np.float32
         )
@@ -94,10 +99,12 @@ class LudoGymEnv(gym.Env):
             i: False for i in range(GameConstants.TOKENS_PER_PLAYER)
         }
 
-        # Internal per-turn state
-        self._pending_agent_dice: Optional[int] = None
-        self._pending_valid_moves: List[Dict] = []
-        self._last_progress_sum: float = 0.0
+        self.obs_builder = ObservationBuilder(self.cfg, self.game, self.agent_color)
+        self.reward_calc = RewardCalculator(self.cfg, self.game, self.agent_color)
+        self.opp_simulator = OpponentSimulator(
+            self.cfg, self.game, self.agent_color, self.move_utils._roll_dice, self.move_utils._make_strategy_context
+        )
+        self.move_utils = MoveUtils(self.cfg, self.game, self.agent_color)
 
     # ----------------------------------------------------------------------------------
     # Gym API
@@ -142,13 +149,13 @@ class LudoGymEnv(gym.Env):
         }
         self._pending_agent_dice = None
         self._pending_valid_moves = []
-        self._last_progress_sum = self._compute_agent_progress_sum()
+        self._last_progress_sum = self.move_utils._compute_agent_progress_sum()
 
         # Advance until it's agent's turn (initial current_player_index is 0 so usually unnecessary)
-        self._ensure_agent_turn()
+        self.opp_simulator._ensure_agent_turn()
         # Roll initial dice for agent decision
-        self._roll_new_agent_dice()
-        obs = self._build_observation()
+        self._pending_agent_dice, self._pending_valid_moves = self.move_utils._roll_new_agent_dice()
+        obs = self.obs_builder._build_observation(self.turns, self._pending_agent_dice)
         self.last_obs = obs
         info = {}
         return obs, info
@@ -172,13 +179,13 @@ class LudoGymEnv(gym.Env):
 
         # Ensure we have a pending dice & valid moves (should always be true except pathological cases)
         if self._pending_agent_dice is None:
-            self._roll_new_agent_dice()
+            self._pending_agent_dice, self._pending_valid_moves = self.move_utils._roll_new_agent_dice()
 
         dice_value = self._pending_agent_dice
         valid_moves = self._pending_valid_moves
 
         # Compute progress baseline
-        progress_before = self._last_progress_sum
+        progress_before = self.move_utils._compute_agent_progress_sum()
 
         # Handle no-move situation
         no_moves_available = len(valid_moves) == 0
@@ -195,9 +202,8 @@ class LudoGymEnv(gym.Env):
             move_res = self.game.execute_move(agent_player, exec_token_id, dice_value)
 
             if move_res.get("captured_tokens"):
-                reward_components.append(
-                    rcfg.capture * len(move_res["captured_tokens"])
-                )
+                capture_reward = self.reward_calc.compute_capture_reward(move_res, reward_components)
+                reward_components.append(capture_reward)
             if move_res.get("token_finished"):
                 reward_components.append(rcfg.finish_token)
             if move_res.get("extra_turn"):
@@ -216,28 +222,23 @@ class LudoGymEnv(gym.Env):
         if not extra_turn and not self.game.game_over:
             # Advance turn for agent (dice consumed)
             self.game.next_turn()
-            self._simulate_opponents(reward_components)
+            self.opp_simulator._simulate_opponents(reward_components)
 
         # Progress shaping (after agent + opponents if any)
-        progress_after = self._compute_agent_progress_sum()
-        delta = progress_after - progress_before
-        if abs(delta) > 1e-9:
-            reward_components.append(delta * rcfg.progress_scale)
+        progress_after = self.move_utils._compute_agent_progress_sum()
+        progress_reward = self.reward_calc.compute_progress_reward(progress_before, progress_after)
+        if progress_reward != 0.0:
+            reward_components.append(progress_reward)
         self._last_progress_sum = progress_after
 
         # Time penalty (light)
         reward_components.append(rcfg.time_penalty)
 
         # Terminal checks
-        terminated = False
-        truncated = False
-        if agent_player.has_won():
-            reward_components.append(rcfg.win)
-            terminated = True
-        elif any(
-            p.has_won() for p in self.game.players if p.color.value != self.agent_color
-        ):
-            reward_components.append(rcfg.lose)
+        opponents = [p for p in self.game.players if p.color.value != self.agent_color]
+        terminal_reward = self.reward_calc.get_terminal_reward(agent_player, opponents)
+        if terminal_reward != 0.0:
+            reward_components.append(terminal_reward)
             terminated = True
 
         self.turns += 1
@@ -248,260 +249,28 @@ class LudoGymEnv(gym.Env):
         # Prepare next agent dice if continuing (extra turn) and not done
         if not terminated and not truncated and not self.game.game_over:
             if extra_turn:
-                self._roll_new_agent_dice()  # agent retains turn
+                self._pending_agent_dice, self._pending_valid_moves = self.move_utils._roll_new_agent_dice()  # agent retains turn
             else:
-                self._ensure_agent_turn()  # move through opponents done above
+                self.opp_simulator._ensure_agent_turn()  # move through opponents done above
                 if not self.game.game_over:
-                    self._roll_new_agent_dice()
+                    self._pending_agent_dice, self._pending_valid_moves = self.move_utils._roll_new_agent_dice()
 
-        obs = self._build_observation()
+        obs = self.obs_builder._build_observation(self.turns, self._pending_agent_dice)
         self.last_obs = obs
         self.done = terminated or truncated
         info = {
             "reward_components": reward_components,
             "dice": self._pending_agent_dice,
             "illegal_action": illegal,
-            "action_mask": self.action_masks(),
+            "action_mask": self.move_utils.action_masks(self._pending_valid_moves),
             "had_extra_turn": extra_turn,
         }
         total_reward = float(sum(reward_components))
         return obs, total_reward, terminated, truncated, info
 
     # ----------------------------------------------------------------------------------
-    # Observation
-    # ----------------------------------------------------------------------------------
-    def _compute_observation_size(self) -> int:
-        base = 0
-        # agent token positions (4)
-        base += 4
-        # opponents token positions (12)
-        base += 12
-        # finished tokens per player (4)
-        base += 4
-        # flags: can_finish, dice_value, progress stats, turn index
-        base += 4  # can_finish, dice_norm, agent_progress, opp_mean_progress
-        if self.cfg.obs_cfg.include_turn_index:
-            base += 1
-        if self.cfg.obs_cfg.include_blocking_count:
-            base += 1
-        return base
-
-    def _normalize_position(self, pos: int) -> float:
-        if pos == GameConstants.HOME_POSITION:
-            return -1.0
-        if pos >= BoardConstants.HOME_COLUMN_START:
-            depth = (
-                pos - BoardConstants.HOME_COLUMN_START
-            ) / GameConstants.HOME_COLUMN_DEPTH_SCALE  # 0..1
-            return (
-                GameConstants.POSITION_NORMALIZATION_FACTOR
-                + depth * GameConstants.POSITION_NORMALIZATION_FACTOR
-            )
-        return (
-            pos / (GameConstants.MAIN_BOARD_SIZE - 1)
-        ) * GameConstants.POSITION_NORMALIZATION_FACTOR  # [0,0.5]
-
-    def _build_observation(self) -> np.ndarray:
-        agent_player = next(
-            p for p in self.game.players if p.color.value == self.agent_color
-        )
-        vec: List[float] = []
-        # agent tokens
-        for t in agent_player.tokens:
-            vec.append(self._normalize_position(t.position))
-        # opponents tokens in fixed global order excluding agent
-        for color in Colors.ALL_COLORS:
-            if color == self.agent_color:
-                continue
-            opp = next(p for p in self.game.players if p.color.value == color)
-            for t in opp.tokens:
-                vec.append(self._normalize_position(t.position))
-        # finished counts
-        for color in Colors.ALL_COLORS:
-            pl = next(p for p in self.game.players if p.color.value == color)
-            vec.append(pl.get_finished_tokens_count() / GameConstants.TOKENS_PER_PLAYER)
-        # can any finish
-        can_finish = 0.0
-        for t in agent_player.tokens:
-            if 0 <= t.position < BoardConstants.HOME_COLUMN_START:
-                remaining = GameConstants.FINISH_POSITION - t.position
-                if remaining <= GameConstants.DICE_MAX:
-                    can_finish = 1.0
-                    break
-        vec.append(can_finish)
-        # dice norm (pending dice for current decision)
-        if self._pending_agent_dice is None:
-            vec.append(0.0)
-        else:
-            vec.append(
-                (self._pending_agent_dice - GameConstants.DICE_NORMALIZATION_MEAN)
-                / GameConstants.DICE_NORMALIZATION_MEAN
-            )
-        # progress stats
-        agent_progress = (
-            agent_player.get_finished_tokens_count() / GameConstants.TOKENS_PER_PLAYER
-        )
-        opp_progresses = []
-        for color in Colors.ALL_COLORS:
-            if color == self.agent_color:
-                continue
-            pl = next(p for p in self.game.players if p.color.value == color)
-            opp_progresses.append(
-                pl.get_finished_tokens_count() / GameConstants.TOKENS_PER_PLAYER
-            )
-        opp_mean = (
-            sum(opp_progresses) / max(1, len(opp_progresses)) if opp_progresses else 0.0
-        )
-        vec.append(agent_progress)
-        vec.append(opp_mean)
-        # turn index scaled
-        if self.cfg.obs_cfg.include_turn_index:
-            vec.append(
-                min(GameConstants.TURN_INDEX_MAX_SCALE, self.turns / self.cfg.max_turns)
-            )
-        # blocking count
-        if self.cfg.obs_cfg.include_blocking_count:
-            blocking_positions = self.game.board.get_blocking_positions(
-                self.agent_color
-            )
-            vec.append(
-                min(
-                    GameConstants.TURN_INDEX_MAX_SCALE,
-                    len(blocking_positions)
-                    / GameConstants.BLOCKING_COUNT_NORMALIZATION,
-                )
-            )  # normalize roughly
-        return np.asarray(vec, dtype=np.float32)
-
-    # ----------------------------------------------------------------------------------
     # Helpers
     # ----------------------------------------------------------------------------------
-    def _roll_dice(self) -> int:
-        # Use core game mechanics (seeding done via global random seed)
-        return self.game.roll_dice()
-
-    def _roll_new_agent_dice(self):
-        self._pending_agent_dice = self._roll_dice()
-        agent_player = next(
-            p for p in self.game.players if p.color.value == self.agent_color
-        )
-        self._pending_valid_moves = self.game.get_valid_moves(
-            agent_player, self._pending_agent_dice
-        )
-
-    def _ensure_agent_turn(self):
-        # Simulate opponents until agent color is current player or game over
-        while (
-            not self.game.game_over
-            and self.game.get_current_player().color.value != self.agent_color
-        ):
-            self._simulate_single_opponent_turn()
-
-    def _simulate_single_opponent_turn(
-        self, reward_components: Optional[List[float]] = None
-    ):
-        """Simulate exactly one opponent player sequence including any extra turns.
-
-        Converted from recursive form to iterative to avoid deep recursion in
-        pathological chains (captures + sixes). Safety-capped.
-        If reward_components provided, apply capture penalties immediately.
-        """
-        current_player = self.game.get_current_player()
-        if current_player.color.value == self.agent_color:
-            return
-        max_chain = GameConstants.MAX_OPPONENT_CHAIN_LENGTH  # generous safety cap
-        chain_count = 0
-        while not self.game.game_over and chain_count < max_chain:
-            dice_value = self._roll_dice()
-            valid_moves = self.game.get_valid_moves(current_player, dice_value)
-            if not valid_moves:
-                # forfeited turn (e.g., 3 sixes or no moves)
-                self.game.next_turn()
-                return
-            try:
-                ctx = self._make_strategy_context(
-                    current_player, dice_value, valid_moves
-                )
-                token_choice = current_player.make_strategic_decision(ctx)
-            except Exception:
-                token_choice = valid_moves[0]["token_id"]
-            move_res = self.game.execute_move(current_player, token_choice, dice_value)
-            # Immediate capture penalty if this move captured agent tokens
-            if reward_components and move_res.get("captured_tokens"):
-                for ct in move_res["captured_tokens"]:
-                    if ct.get("player_color") == self.agent_color:
-                        reward_components.append(self.cfg.reward_cfg.got_captured)
-            if not move_res.get("extra_turn") or self.game.game_over:
-                if not self.game.game_over:
-                    self.game.next_turn()
-                return
-            chain_count += 1
-        # Safety: force turn end if chain exceeded
-        if not self.game.game_over and self.game.get_current_player() is current_player:
-            self.game.next_turn()
-
-    def _simulate_opponents(self, reward_components: List[float]):
-        # Simulate opponents in order until agent's turn or game over.
-        # Capture penalties are handled inside _simulate_single_opponent_turn now.
-        while (
-            not self.game.game_over
-            and self.game.get_current_player().color.value != self.agent_color
-        ):
-            self._simulate_single_opponent_turn(reward_components)
-
-    def _snapshot_agent_tokens(self) -> List[int]:
-        player = next(p for p in self.game.players if p.color.value == self.agent_color)
-        return [t.position for t in player.tokens]
-
-    def _compute_agent_progress_sum(self) -> float:
-        player = next(p for p in self.game.players if p.color.value == self.agent_color)
-        total = 0.0
-        for t in player.tokens:
-            if t.position == -1:
-                continue
-            if 0 <= t.position < GameConstants.MAIN_BOARD_SIZE:
-                total += t.position / float(
-                    GameConstants.MAIN_BOARD_SIZE + GameConstants.HOME_COLUMN_SIZE
-                )
-            elif t.position >= BoardConstants.HOME_COLUMN_START:
-                # Home column progress: 52..58 mapped
-                offset = (
-                    t.position - BoardConstants.HOME_COLUMN_START
-                ) + GameConstants.MAIN_BOARD_SIZE
-                total += offset / float(
-                    GameConstants.MAIN_BOARD_SIZE + GameConstants.HOME_COLUMN_SIZE
-                )
-        return total
-
-    # Action masking (for algorithms that support it)
-    def action_masks(self) -> np.ndarray:
-        mask = np.zeros(GameConstants.TOKENS_PER_PLAYER, dtype=np.int8)
-        if self._pending_valid_moves:
-            valid_ids = {m["token_id"] for m in self._pending_valid_moves}
-            for i in range(GameConstants.TOKENS_PER_PLAYER):
-                if i in valid_ids:
-                    mask[i] = 1
-        return mask
-
-    def _make_strategy_context(
-        self, player: Player, dice_value: int, valid_moves: List[Dict]
-    ):
-        # Basic context bridging existing strategies expecting a structure similar to tournaments
-        board_state = self.game.board.get_board_state_for_ai(player)
-        opponents = []
-        for p in self.game.players:
-            if p is player:
-                continue
-            opponents.append(p.get_game_state())
-        ctx = {
-            "player_state": player.get_game_state(),
-            "board": board_state,
-            "valid_moves": valid_moves,
-            "dice_value": dice_value,
-            "opponents": opponents,
-        }
-        return ctx
-
     def render(self):  # minimal
         print(f"Turn {self.turns} agent_color={self.agent_color}")
