+++ ludo_rl/envs/simulators/__init__.py
@@ -1,0 +1,1 @@
+# Simulators package
