+++ ludo_rl/envs/utils/__init__.py
@@ -1,0 +1,1 @@
+# Utils package
